add universal afct implementation based on pytorch
